+++ EasyAccess/Features/CollectOutputs.cs
@@ -84,6 +84,7 @@ internal class CollectOutputs : Feature
                 name = "Collect Outputs",
                 hoverText = I18n.Button_CollectOutputs_Name(),
                 baseScale = 2f,
+                scale = 2f,
             },
             ComponentArea.Right);
     }